app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
databricks-sdk>=0.30.0
pydantic>=2.0.0
python-multipart>=0.0.9
